http://issues.apache.org/activemq/browse/AMQCPP-30
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@508988 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/connector/openwire/commands/DestinationInfo.cpp
@@ -124,7 +124,7 @@ std::string DestinationInfo::toString() const {
     } else {
         stream << "   Object is NULL" << std::endl;
     }
-    stream << " Value of OperationType = " << this->getOperationType() << std::endl;
+    stream << " Value of OperationType = " << (int)this->getOperationType() << std::endl;
     stream << " Value of Timeout = " << this->getTimeout() << std::endl;
     for( size_t ibrokerPath = 0; ibrokerPath < this->getBrokerPath().size(); ++ibrokerPath ) {
         stream << " Value of BrokerPath[" << ibrokerPath << "] is Below:" << std::endl;
